global refactoring mechanizm retrieving roles, files, auth providers. removing concrete methods, added method to find by id. refactorgin signup and login controllers, added addition methods to user service - create user, create facebook user and etc.

+++ src/main/java/com/smartestgift/dao/FileDAO.java
@@ -6,5 +6,5 @@ import com.smartestgift.dao.model.File;
  * Created by dikkini on 19.11.13.
  * Email: dikkini@gmail.com
  */
-public interface FileDAO extends Repository<File, Long> {
+public interface FileDAO extends Repository<File, Integer> {
 }
